requirements/base.txt
Django==3.2.6
Pillow==8.3.1
djangorestframework==3.12.4
django-cors-headers==3.8.0
djangorestframework-simplejwt==4.8.0